check: switch back to initial path after repo commands
We switch to CACHEDIR to store the repo init/sync output but we never
switch back to the original directory (OLDPWD).

This causes issues when we call check() in unit tests and expect to be
in the default project directory.

Restore back to PWD after we are done running check()

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>