common: add jobs source configuration option
By default, repo sync will run as many jobs as the amount of CPU cores on the
host machine.

This is fine for smaller development machines, but for some pipelines might want
to adjust this. For example, when rate-limiting is applied.

Add a 'jobs' source configuration field to do so.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>